% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webp.R
\name{read_webp}
\alias{read_webp}
\title{Webp image format}
\usage{
read_webp(buf, numeric = TRUE)
}
\description{
Read and write webp images into a raw bitmap array.
}
\examples{
# Convert to webp
library(png)
img <- readPNG(system.file("img", "Rlogo.png", package="png"))
write_webp(img, "rlogo.webp")
browseURL("rlogo.webp")
rm(img)

# Convert from webp
library(jpeg)
img <- read_webp("rlogo.webp")
writeJPEG(img, "rlogo.jpeg")
browseURL("rlogo.jpeg")
}

